BUG FIX: Fix swapchain recreation failing
Bug fix related to swapchain artifacts and not updating extent.

Swapchain recreation chooses a new extent based on the Physical Devices knowledge of the current surface.

Incorrectly I cached this information and never prompted the Physical Device for updated data. This meant on a screen resize the extent would stay the same. Resulting in problematic artifacts.

Now GetDeviceSwapchainSupport() is a public function in our PhysicalDevice.cpp and SwapchainSupportDetails are not cached anywhere. This is called at the beginning of each swapchain creation, fixing the artifacts and my sleep schedule.