Optimize Vertex Buffer Creation on GPU
Instead of creating a VkBuffer on the GPU and writing to it from the CPU, we create two buffers.
* One opens the CPU to write data onto the GPU.
* The second is strictly for the GPU.

The first is a temp buffer, whose data is copied to the GPU only buffer. This allows us to end with a buffer only on the GPU. Reducing read times as CPU coherency is not required.