Pull instance data out of dynamic UBO into static SSBO
Instance data like texture index was being added to the GPU every frame along with our camera position. This was not necessary as instance data currently only changes on world load.

So in the new solution I created a SSBO that is initialized only when a model is loaded.

This comes with the added benefit of having a unbound instance count, and a simpler UBO call scheme.

I have also realized some of my wrappings around buffer creation are too narrow and might do more harm than good, so for future refactoring I may iterate on them.