BUG FIX: Fix mesh point collapse
Vertices coming out of the Vertex Shader were returning 0,0,0,0. Which means one of the matrices must be a 4x4 of 0's.

To see which matrix is responsible, we can remove each matrix from the vertex shader. After removing the model matrix, we can see the effect vanish, leading us to investigate the model matrix.

RenderDoc allows us to see the instance_data array as the GPU sees it, which stores each of our model matrices. It matches perfectly to the number of instances, and none are a 4x4 matrix of 0s.

So the data on the GPU is correct, but the information the shader is getting is wrong.

So finally, in a grand mystery this points us to the places we tell the GPU where to look for the instance data. The descriptor creation was correct, same with the binding, but something was off with the VkDrawIndexedIndirectCommand creation.

Inside of VkDrawIndexedIndirectCommand, the first_instance parameter tells the GPU where to start rendering the given instances of the draw call. In later draw calls, this value was over 1 Million, while we only had ~40k mesh instances. Vulkan defaults out of bound array calls to a 4x4 of 0's, leading to our error.

Fixing this variable in Renderer.Cpp, the error is fixed.